done feature

+++ app/src/main/java/com/challenge/foodappchallenge3/data/network/api/model/order/OrderRequest.kt
@@ -9,7 +9,7 @@ data class OrderRequest(
     val username: String?,
 
     @SerializedName("total")
-    val total: Int?,
+    val total: Double?,
 
     @SerializedName("orders")
     val orders: List<OrderItemRequest>?
